Σ extract

Extract Bayesian estimates of Σ from the function niw.post() in the package nicheROVER.

Usage

sigma_extract(data, isotope_a = NULL, isotope_b = NULL, format = "wide")

Arguments

data

a list created by the function niw.post() in the package nicheROVER

isotope_a

character string to supply for the first isotope used in niw.post(). Defaults to "d15n".

isotope_b

characterstring to supply for the second isotope used in niw.post(). Defaults to "d13c".

format

format a character that decided whether the returned object is in long or wide format. Default is "wide", with the alternative supplied being "long".

Value

Returns a tibble of extracted estimates of Σ created by the function niw.post() in the package nicheROVER. The tibble will contain five columns in the following order, metric, id, sample_name, isotope, sample_number, and the posterior sample for Σ (i.e., post_sample).

See also

Examples

sigma_extract(
data = niw_fish_post
)
#> # A tibble: 8,000 × 6
#>    metric sample_name isotope sample_number  d15n  d13c
#>    <chr>  <chr>       <chr>   <chr>         <dbl> <dbl>
#>  1 sigma  ARCS        d15n    1             0.501 0.216
#>  2 sigma  ARCS        d13c    1             0.216 1.07 
#>  3 sigma  ARCS        d15n    2             0.577 0.206
#>  4 sigma  ARCS        d13c    2             0.206 0.844
#>  5 sigma  ARCS        d15n    3             0.725 0.449
#>  6 sigma  ARCS        d13c    3             0.449 1.48 
#>  7 sigma  ARCS        d15n    4             0.888 0.324
#>  8 sigma  ARCS        d13c    4             0.324 1.07 
#>  9 sigma  ARCS        d15n    5             0.681 0.266
#> 10 sigma  ARCS        d13c    5             0.266 0.970
#> # ℹ 7,990 more rows
# ---- below will be turned into vignette when finished ----
if (FALSE) {
# load packages
{
  library(dplyr)
  library(janitor)
  library(nicheROVER)
  library(nichetools)
  library(purrr)
  }

# ---- To understand how niw_fish_post is being created ----

# get fish data frame, and remove sulfur for the example
df <- fish %>%
    janitor::clean_names()

# create number of samples used in nicheROVER
nsample <- 1000

# split the data frame by species and select isotopes of interest
df_split <- df %>%
  split(.$species) %>%
  map(~ select(., d15n, d13c))

# extract the names of each list to name each object in list
df_names <- df %>%
   group_by(species) %>%
   group_keys() %>%
   ungroup() %>%
   mutate(
     id = 1:nrow(.)
     )

# name each object in list
names(df_split) <- df_names$species

# create niw posterior samples
niw_fish_post <- df_split %>%
  map(~niw.post(nsample = nsample, X = .))

# we can then use sigma_extract() to extract sigma from niw_fish_post
df_sigma <- sigma_extract(
  data = niw_fish_post
  )

# --- to use with `niche_ellipse()` we need to make into wide format ----
df_sigma_wide <- df_sigma |>
tidyr::pivot_wider(names_from = id,
            values_from = post_sample)
}