extract Layman metrics

Extract Bayesian estimates for the following six layman metrics, \(\delta^{13}\)C range, \(\delta^{15}\)N range, total area (TA), distance to centroid (CD), distance to the nearest neighbour (NND), and the standard deviation of the distance to the nearest neighbour (SDNND) from data objects created by SIBER. To learn more about the following metrics please review Layman et al. (2008).

Usage

extract_layman(data, community_df = NULL, data_format = NULL)

Arguments

data

a list created by the function bayesianLayman() from the package SIBER.

community_df

a two column data frame. One of the columns has to be named community and the data in the column will be numerics as a character string(e.g., "1", "2", "3"). This is the order of the community names and will be used to join the actual community names to the correct data. These are the same class and values required by the function, createSiberObject() from SIBER. The second column contains the actual names of the communities that the user is working with (e.g., "region").

data_format

a character string that decides whether the returned object is in long or wide format. Default is "long", with the alternative supplied being "wide".

Value

A tibble containing four rows when data_format is set to its default which is long. These four rows are the following, community, the_name_of_the_communities, metric and post_est.

See also

Examples

library(nichetools)
library(SIBER)

# ---- bring in SIBER demo data ----
# uncomenet to use
# str(demo.siber.data)

# ---- create coummnity names data frame ----
# uncomment to use
# str(demo.siber.data.2)

demo.siber.data.2$community_names <- as.factor(demo.siber.data.2$community)

demo.siber.data.2$community <- as.numeric(demo.siber.data.2$community_names) |>
as.character()
c_names <- demo.siber.data.2 |>
dplyr::distinct(community, community_names)

# ---- create the siber object ----
siber.example <- createSiberObject(demo.siber.data)

# ---- view Baysain estimates of mu and sigma produced by SIBER ---
# uncomment to use
# str(post_sam_siber)

# ---- extract posterior estimates of mu -----

mu_post <- extractPosteriorMeans(siber.example, post_sam_siber)

# ---- Bayesian estimates of layman metrics using SIBER ----

layman_b <- bayesianLayman(mu.post = mu_post)

# ---- use nichetools to extract Baysian estimats of Layman metrics ----

layman_be <- extract_layman(layman_b, community_df = c_names)

layman_be
#> # A tibble: 48,000 × 4
#>    community community_names metric   post_est
#>    <chr>     <fct>           <chr>       <dbl>
#>  1 1         dublin          dY_range     6.26
#>  2 1         dublin          dX_range     7.92
#>  3 1         dublin          TA          11.2 
#>  4 1         dublin          CD           4.12
#>  5 1         dublin          NND          4.67
#>  6 1         dublin          SDNND        3.69
#>  7 1         dublin          dY_range     6.27
#>  8 1         dublin          dX_range     8.22
#>  9 1         dublin          TA          13.5 
#> 10 1         dublin          CD           4.20
#> # ℹ 47,990 more rows